convert everything to decimal and minor tweaking

+++ PoloniexClient/Bittrex/BittrexAccountBalanceInfo.cs
@@ -7,9 +7,9 @@ using System.Threading.Tasks;
 namespace CryptoMarketClient.Bittrex {
     public class BittrexAccountBalanceInfo {
         public string Currency { get; set; }
-        public double Balance { get; set; }
-        public double Available { get; set; }
-        public double Pending { get; set; }
+        public decimal Balance { get; set; }
+        public decimal Available { get; set; }
+        public decimal Pending { get; set; }
         public string CryptoAddress { get; set; }
         public bool Requested { get; set; }
         public string Uuid { get; set; }
